Move sync to own function
